requirements.txt
Django>=5.0,<6.0
djangorestframework
celery
redis
psycopg2-binary
django-htmx
python-decouple
gunicorn
whitenoise
dj-database-url
lxml
pandas
openpyxl
django-htmx
django-cors-headers
whitenoise
psycopg2-binary
Pillow>=10.0.0
requests

